Add country and company attributes to Person
The add command currently uses 'address' as an attribute of Person.
This attribute name does not fit the intended usage.

Let's rename the 'address' attribute to 'country' throughout the
codebase. In addition, let's introduce a new 'company' attribute
that shares the same input constraints as 'country'.

Test cases have also been updated and added to ensure the
correctness of the modified and newly introduced attributes.

These changes support the direction of the model for our
target audience.