Add OS time display to Person
PersonCard does not display local time zones which is required
by the user to ensure he contacts clients at appropriate times.

Let's add in a Time class which is also included in a Person,
this way we can access the local time of the person to be displayed
through PersonCard.

Test cases have been added to ensure time displayed is accurate and
correct.